Don't use tasklist.exe anymore to get pids from process names, instead use our own python code calling into ntdll.
Extend GETPIDS support so that we acquire a list of processes to be hidden prior to the start of the malware.  The current list of names (from supahal) is mostly irrelevant, it's unlikely wireshark/filemon/etc will be running in anyone's sandbox prior to the malware starting, and malware generally checks for the vm user tools in other ways as well.  We can't move this to after the malware starts or we risk both a trivial detection measure and a way to avoid analysis.