do not exhaust local sockets when no network interface is available
It would be possible that the network interface was not yet available
due to which the socket.create_connection() function throws an exception
right away. Following the infinite loop this would exhaust all the
sockets which in turn resulted in not being able to reach the host and
thus a completely failed analysis.

(cherry picked from commit 08e2877e474d960443a6d4127a63501178726052)

Conflicts:
	analyzer/android/lib/common/results.py